move: global/config 디렉토리 세분화
