--- src/main/java/com/kgu/studywithme/global/BaseDomainEvent.java
@@ -1,10 +1,0 @@
-package com.kgu.studywithme.global;
-
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-public abstract class BaseDomainEvent {
-    private final LocalDateTime eventPublishedDate = LocalDateTime.now();
-}
