+++ src/test/java/com/kgu/studywithme/common/config/TestWebBeanConfiguration.java
@@ -1,7 +1,7 @@
 package com.kgu.studywithme.common.config;
 
 import com.kgu.studywithme.auth.utils.TokenResponseWriter;
-import com.kgu.studywithme.global.config.CorsProperties;
+import com.kgu.studywithme.global.config.web.CorsProperties;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
